Update switch.py
Love this idea for integration. I have a version which I think you have a fork of that specifically does this integration with the older CL version. I have not had time to re-integrate on the newer CL version. I have not tested it yet, but I added logic to automatically grab the Hue info from the HA Hue plugin. This should get the API_KEY (I.e., username) and the IP Address of the bridge. So if a user has added the Hue integration into HA, this will auto-add that info.